+++ brala/network/packets/util.d
@@ -129,6 +129,6 @@ mixin template Packet(ubyte id_, Vars...) {
     mixin(inject_data());
 }
 
-vec3i coords_from_j(size_t j, uint y) {
+vec3i coords_from_j(uint j, uint y) {
     return vec3i(j & 0x0F, y*16 + (j >> 8), (j & 0xF0) >> 4);
 }
